replaces listActiveGames with GameListing component
